+++ tests/scheduler/backends/qblox/operation_handling/empty_qasm_program.py
@@ -9,4 +9,8 @@ from quantify_scheduler.backends.qblox.register_manager import RegisterManager
 @pytest.fixture(name="empty_qasm_program_qcm")
 def fixture_empty_qasm_program():
     """Empty QASMProgram object."""
-    yield QASMProgram(QcmModule.static_hw_properties, RegisterManager())
+    yield QASMProgram(
+        static_hw_properties=QcmModule.static_hw_properties,
+        register_manager=RegisterManager(),
+        align_fields=True,
+    )