+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 - Compilation - Add a `scheduling_strategy` parameter to `QuantumDevice` and `DeviceCompilationConfig` classes to enable new strategies for calculating absolute timing in `compilation.determine_absolute_timing`  (!736)
 - Security - Add `check=True` flag to all subprocess calls (see also Ruff rule PLW1510). (!767)
+- Qblox backend - Make QASM fields aligning optional, disable by default (!741)
+- Documentation - Color change for code block emphasized lines (!741)
 
 ## 0.16.0 (2023-08-17)